Move smallrye-config-source-file-system packages to io.smallrye.config.source.file
