+++ sources/file-system/src/test/resources/io/smallrye/config/source/file/configDir/myKey1
